pyproject.toml
[tool.poetry]
name = "transcriber"
version = "0.1.0"
description = "CLI-tool for transcription and subtitles(faster-whisper + ffmpeg)"
authors = ["a1deas <roland.brannon.am@gmail.com>"]
readme = "README.md"
packages = [{ include = "transcriber" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
faster-whisper = "^1.0.0"
ffmpeg-python = "^0.2.0"
typer = "^0.12.0"
rich = "^13.0.0"
click = ">=8.1.7,<8.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[tool.poetry.scripts]
transcriber = "transcriber.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"